refactor: [017-FIX-WRONG-TERM] 상품 DTO 패키지명 Production -> Product 오타 수정
