+++ src/main/java/com/myecommerce/MyECommerce/dto/product/RequestProductionDto.java
@@ -1,4 +1,4 @@
-package com.myecommerce.MyECommerce.dto.production;
+package com.myecommerce.MyECommerce.dto.product;
 
 import com.myecommerce.MyECommerce.type.ProductCategoryType;
 import com.myecommerce.MyECommerce.validation.EnumValid;
