+++ src/main/java/com/myecommerce/MyECommerce/dto/product/RequestProductionOptionDto.java
@@ -1,4 +1,4 @@
-package com.myecommerce.MyECommerce.dto.production;
+package com.myecommerce.MyECommerce.dto.product;
 
 import jakarta.validation.constraints.*;
 import lombok.*;
@@ -10,9 +10,7 @@ import java.math.BigDecimal;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class RequestModifyProductionOptionDto {
-
-    private Long id;
+public class RequestProductionOptionDto {
 
     @NotBlank(message = "{validation.product.option.code.not.blank}")
     @Size(min=1, max=100, message = "{validation.product.option.code.size}")
