+++ src/main/java/com/myecommerce/MyECommerce/dto/product/ResponseSearchDetailProductionDto.java
@@ -1,9 +1,10 @@
-package com.myecommerce.MyECommerce.dto.production;
+package com.myecommerce.MyECommerce.dto.product;
 
 import com.myecommerce.MyECommerce.type.ProductCategoryType;
 import com.myecommerce.MyECommerce.type.ProductSaleStatusType;
 import lombok.*;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @Getter
@@ -11,12 +12,30 @@ import java.util.List;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ServiceProductionDto {
+public class ResponseSearchDetailProductionDto {
+
     private Long id;
+
+    private Long seller;
+
     private String code;
+
     private String name;
+
     private ProductCategoryType category;
+
     private String description;
+
     private ProductSaleStatusType saleStatus;
-    private List<ServiceProductionOptionDto> options;
+
+    private Long createId;
+
+    private LocalDateTime createDt;
+
+    private Long updateId;
+
+    private LocalDateTime updateDt;
+
+    private List<ResponseProductionOptionDto> options;
+
 }
