+++ src/main/java/com/myecommerce/MyECommerce/dto/product/ServiceProductionDto.java
@@ -1,10 +1,9 @@
-package com.myecommerce.MyECommerce.dto.production;
+package com.myecommerce.MyECommerce.dto.product;
 
 import com.myecommerce.MyECommerce.type.ProductCategoryType;
 import com.myecommerce.MyECommerce.type.ProductSaleStatusType;
 import lombok.*;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Getter
@@ -12,30 +11,12 @@ import java.util.List;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ResponseSearchDetailProductionDto {
-
+public class ServiceProductionDto {
     private Long id;
-
-    private Long seller;
-
     private String code;
-
     private String name;
-
     private ProductCategoryType category;
-
     private String description;
-
     private ProductSaleStatusType saleStatus;
-
-    private Long createId;
-
-    private LocalDateTime createDt;
-
-    private Long updateId;
-
-    private LocalDateTime updateDt;
-
-    private List<ResponseProductionOptionDto> options;
-
+    private List<ServiceProductionOptionDto> options;
 }
