+++ src/main/java/com/myecommerce/MyECommerce/dto/product/ServiceProductionOptionDto.java
@@ -1,4 +1,4 @@
-package com.myecommerce.MyECommerce.dto.production;
+package com.myecommerce.MyECommerce.dto.product;
 
 import lombok.*;
