+++ src/test/java/com/myecommerce/MyECommerce/controller/ProductionControllerTest.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.myecommerce.MyECommerce.config.JwtAuthenticationProvider;
 import com.myecommerce.MyECommerce.config.SecurityConfig;
 import com.myecommerce.MyECommerce.controller.config.TestSecurityConfig;
-import com.myecommerce.MyECommerce.dto.production.RequestProductionDto;
-import com.myecommerce.MyECommerce.dto.production.RequestProductionOptionDto;
-import com.myecommerce.MyECommerce.dto.production.ResponseProductionDto;
+import com.myecommerce.MyECommerce.dto.product.RequestProductionDto;
+import com.myecommerce.MyECommerce.dto.product.RequestProductionOptionDto;
+import com.myecommerce.MyECommerce.dto.product.ResponseProductionDto;
 import com.myecommerce.MyECommerce.entity.member.Member;
 import com.myecommerce.MyECommerce.entity.member.MemberAuthority;
 import com.myecommerce.MyECommerce.exception.ProductException;
